Add constraints to microservices venv
Constraints file is taken from requirements repo URL, pip in venv is
configured to always use it and never use cache dir (it's a very common
option in our Dockerfiles).

Note that the plan was to get URL from config but there's currently no
good place in configs for it.

Also removed --system-site-packages option to virtualenv as it defeats
the purpose of constraints.

Change-Id: I03a19b423b6617539d91fb29f44b40271849174f