Use constraints when install virtualenv
Without this we get too fresh virtualenv while we have packages versions
conflicting with it in constraints.txt. Also disable downloading random
(fresh) versions of packages by virtualenv and remove --upgrade from
pip. setuptools package is not mentioned in constraints.txt because it
is expected to be installed at proper version by virtualenv, so we
shouldn't upgrade it to later versions.

Right now we end up installing latest virtualenv 15.1.0 with
setuptools 34.0.0 which depends on packaging>=16.8 while for Newton we
have packaging===16.7.0 in constraints.txt

Change-Id: I07ea9e2cf83dd0c1bf330556ad0aee7139dd1ef0